completed #51 and fixed issues with setpwd

+++ include/drivers/rng/virtio_rng.h
@@ -3,5 +3,5 @@
 #include <drivers/virtio/virtio.h>
 
 int virtio_rng_init();
-int virtio_rng_read(u8* buf, usize len);
+usize virtio_rng_read(u8* buf, usize len);
 bool virtio_rng_available();
